fix: Component name refactoring
