+++ src/Footer/index.jsx
@@ -1,12 +1,11 @@
 import React from 'react';
-import styles from './Footer.module.css';
+import styles from './index.module.css';
 function Footer(props) {
-
     return (
         <footer >
             <div className={styles.Footer}>
-            <p>Anton Zhekov SoftUni project &#169; 2020</p>
-            </div>            
+                <p>Anton Zhekov SoftUni project &#169; 2020</p>
+            </div>
         </footer>);
 }
