--- src/contexts/UserContext.js
@@ -1,5 +1,0 @@
-import React from "react";
-
-const UserContext = React.createContext({name: "Anton"});
-
-export default UserContext
